+++ src/extension.ts
@@ -1,22 +1,10 @@
-import TelemetryReporter from '@vscode/extension-telemetry';
 import * as vscode from 'vscode';
-
-const brand = `HashiCorp HCL`;
-const outputChannel = vscode.window.createOutputChannel(brand);
-
-let reporter: TelemetryReporter;
+import { activateExtension, deactivateExtension } from './common';
 
 export async function activate(context: vscode.ExtensionContext): Promise<void> {
-  const manifest = context.extension.packageJSON;
-  reporter = new TelemetryReporter(context.extension.id, manifest.version, manifest.appInsightsKey);
-  context.subscriptions.push(reporter);
-  context.subscriptions.push(outputChannel);
-
-  reporter.sendTelemetryEvent('startExtension');
-  outputChannel.appendLine(`Started: HCL ${vscode.env.appHost}`);
+  activateExtension(context);
 }
 
 export async function deactivate(): Promise<void> {
-  reporter.sendTelemetryEvent('stopExtension');
-  outputChannel.appendLine(`Stopped: HCL ${vscode.env.appHost}`);
+  deactivateExtension();
 }
